+++ src/lib/filters/filter-brightness.ts
@@ -3,6 +3,7 @@ import { uuid } from '../core';
 
 export function filterBrightness(selection: Selection<SVGFilterElement>, brightness: number): void {
   selection
+    .classed('filter-brightness', true)
     .attr('id', uuid())
     .append('feComponentTransfer')
     .call((s) =>